# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Note App API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/notes/": {
            "get": {
                "tags": [
                    "Note"
                ],
                "summary": "Получение заметок",
                "operationId": "3aa8659d31289597f3ef0245baf47b2d",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "created_at": {
                                                    "type": "timestamp",
                                                    "example": "2023-11-28 17:50:37"
                                                },
                                                "updated_at": {
                                                    "type": "timestamp",
                                                    "example": null
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Some title"
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "fields": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "created_at": {
                                                                "type": "timestamp",
                                                                "example": "2023-11-28 17:50:37"
                                                            },
                                                            "updated_at": {
                                                                "type": "timestamp",
                                                                "example": null
                                                            },
                                                            "title": {
                                                                "type": "string",
                                                                "example": "Worker"
                                                            },
                                                            "type": {
                                                                "type": "enum('string', 'integer', 'float', 'boolean')",
                                                                "example": "string"
                                                            },
                                                            "value": {
                                                                "type": "string",
                                                                "example": "Ivan Ivanov"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Note"
                ],
                "summary": "Создание заметки",
                "operationId": "7e23f5a8817b1ec6fae2689eff1f4227",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "Work"
                                            },
                                            "fields": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "title": {
                                                            "type": "string",
                                                            "example": "Worker"
                                                        },
                                                        "type": {
                                                            "type": "enum('string', 'integer', 'float', 'boolean')",
                                                            "example": "string"
                                                        },
                                                        "value": {
                                                            "type": "string",
                                                            "example": "Ivan Ivanov"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "title": {
                                                    "type": "string",
                                                    "example": "Some title"
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "New note created."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/notes/{id}": {
            "put": {
                "tags": [
                    "Note"
                ],
                "summary": "Редактирование заметки",
                "operationId": "c7e6224fadd6d82b81ded7615279a15c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID заметки",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "Work"
                                            },
                                            "fields": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "title": {
                                                            "type": "string",
                                                            "example": "Worker"
                                                        },
                                                        "type": {
                                                            "type": "string",
                                                            "example": "string"
                                                        },
                                                        "value": {
                                                            "type": "string",
                                                            "example": "Ivan Ivanov"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Note is updated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Note"
                ],
                "summary": "Удаление заметки",
                "operationId": "231822d83675b5b9446d8b0f6f4d3959",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID заметки",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Note is deleted."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}